requirements.txt
pyTelegramBotAPI==4.12.0
python-dotenv
psycopg2-binary==2.9.9
